refactor: split helpers file
