+++ src/components/ds/InputDateGroup.tsx
@@ -6,7 +6,7 @@ import { Box, Flex, Input, FormControl, FormErrorMessage, FormLabel } from "@cha
 import { Field } from "react-final-form"
 
 import { isFieldHasError, validateDate } from "../../utils/formHelpers"
-import { dateToString, parseDate } from "../../utils/helpers"
+import { dateToString, parseDate } from "../../utils/date"
 
 registerLocale("fr", fr)
