+++ src/utils/calculsEgaProIndicateurDeux.tsx
@@ -1,6 +1,6 @@
 import { AppState, CategorieSocioPro, GroupeEffectif, GroupeIndicateurDeux } from "../globals"
 
-import { roundDecimal } from "./helpers"
+import { roundDecimal } from "./number"
 
 import {
   calculEcartsPonderesParGroupe,
