+++ src/utils/calculsEgaProIndicateurQuatre.tsx
@@ -1,6 +1,6 @@
 import { AppState } from "../globals"
 
-import { roundDecimal } from "./helpers"
+import { roundDecimal } from "./number"
 
 //////////////////
 // INDICATEUR 4 //
