+++ src/utils/calculsEgaProIndicateurTrois.tsx
@@ -1,6 +1,6 @@
 import { AppState, CategorieSocioPro, GroupeEffectif, GroupeIndicateurTrois } from "../globals"
 
-import { roundDecimal } from "./helpers"
+import { roundDecimal } from "./number"
 
 import {
   calculEcartsPonderesParGroupe,
