+++ src/views/Declaration/DeclarationForm.tsx
@@ -6,7 +6,8 @@ import { AppState, FormState, ActionDeclarationData } from "../../globals"
 import { resendReceipt } from "../../utils/api"
 import { displayMetaErrors } from "../../utils/form-error-helpers"
 import { isFormValid, parseBooleanFormValue, parseBooleanStateValue, required } from "../../utils/formHelpers"
-import { logToSentry, parseDate } from "../../utils/helpers"
+import { parseDate } from "../../utils/date"
+import { logToSentry } from "../../utils/sentry"
 
 import InputDateGroup from "../../components/ds/InputDateGroup"
 import ActionBar from "../../components/ActionBar"
